bug fix (MapFeirasScreen): correção para fechamento repentino quando localizacao fica desabilitada.
